Final CSV refactor
