+++ CLI/commands/accredit.js
@@ -1,79 +1,35 @@
 var common = require('./common/common_functions');
-var csv = require('./common/csv_sync');
-var contracts = require('./helpers/contract_addresses');
+var csv_shared = require('./common/csv_shared');
 var abis = require('./helpers/contract_abis');
 var BigNumber = require('bignumber.js');
 
 let distribData = new Array();
-let allocData = new Array();
 let fullFileData = new Array();
 let badData = new Array();
 
-let BATCH_SIZE = 70;
 let securityToken;
 
 async function startScript(tokenSymbol, batchSize) {
-  if (batchSize) {
-    BATCH_SIZE = batchSize;
-  }
+  securityToken = await csv_shared.start(tokenSymbol, batchSize);
 
-  //common.logAsciiBull();
-
-  let STAddress = await checkST(tokenSymbol);
-  securityToken = new web3.eth.Contract(abis.securityToken(), STAddress);
+  let result_processing = await csv_shared.read('./CLI/data/accredited_data.csv', accredit_processing);
+  distribData = result_processing.distribData;
+  fullFileData = result_processing.fullFileData;
+  badData = result_processing.badData;
   
-  await readCsv();
+  await saveInBlockchain();
 };
 
-async function checkST(tokenSymbol) {
-  let securityTokenRegistry = await STConnect();
+function accredit_processing(csv_line) {
+  let isAddress = web3.utils.isAddress(csv_line[0]);
+  let isAccredited = (typeof JSON.parse(csv_line[1].toLowerCase())) == "boolean" ? JSON.parse(csv_line[1].toLowerCase()) : "not-valid";
 
-  return await securityTokenRegistry.methods.getSecurityTokenAddress(tokenSymbol).call({}, function (error, result) {
-    if (result != "0x0000000000000000000000000000000000000000") {
-      return result
-    } else {
-      console.log("Token doesn't exist")
-      process.exit(0)
-    }
-  });
-}
-
-async function STConnect() {
-  try {
-    let STRegistryAddress = await contracts.securityTokenRegistry();
-    let STRegistry = new web3.eth.Contract(abis.securityTokenRegistry(), STRegistryAddress);
-    return STRegistry;
-  } catch (err) {
-    console.log("There was a problem getting the contracts. Make sure they are deployed to the selected network.");
-    process.exit(0);
+  if (isAddress &&
+    (isAccredited != "not-valid")) {
+    return [true, new Array(web3.utils.toChecksumAddress(csv_line[0]), isAccredited)]
+  } else {
+    return [false, new Array(csv_line[0], isAccredited)]
   }
-}
-
-async function readCsv() {
-  var CSV_STRING = csv('./CLI/data/accredited_data.csv');
-  let i = 0;
-
-  CSV_STRING.forEach(line => {
-    let data_processed = accredit_processing(line);
-    fullFileData.push(data_processed[1]);
-
-    if (data_processed[0]) {
-      allocData.push(data_processed[1]);
-      i++;
-      if (i >= BATCH_SIZE) {
-        distribData.push(allocData);
-        allocData = [];
-        i = 0;
-      }
-    } else {
-      badData.push(data_processed[1]);
-    }
-  });
-
-  distribData.push(allocData);
-  allocData = [];
-
-  await saveInBlockchain();
 }
 
 async function saveInBlockchain() {
@@ -129,22 +85,6 @@ async function saveInBlockchain() {
   }
 
   return;
-}
-
-function accredit_processing(csv_line) {
-  let isAddress = web3.utils.isAddress(csv_line[0]);
-  let isAccredited = (typeof JSON.parse(csv_line[1].toLowerCase())) == "boolean" ? JSON.parse(csv_line[1].toLowerCase()) : "not-valid";
-
-  if (isAddress &&
-    (isAccredited != "not-valid")) {
-
-    return [true, new Array(web3.utils.toChecksumAddress(csv_line[0]), isAccredited)]
-
-  } else {
-
-    return [false, new Array(csv_line[0], isAccredited)]
-  
-  }
 }
 
 module.exports = {
